requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-fsm==3.0.0
django-model-utils==5.0.0
django-notifications-hq==1.8.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jsonfield==3.1.0
msgpack==1.1.0
pillow==11.1.0
PyJWT==2.10.1
pytz==2025.1
redis==5.2.1
sqlparse==0.5.3
swapper==1.4.0
